requirements.txt
python-dotenv
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
fastapi-mcp